requirements.txt
torch
transformers
peft
streamlit
pandas
numpy
scikit-learn
Pillow
joblib
accelerate
